pyproject.toml
[project]
name = "rickbot"
version = "0.2"
description = "A chatbot based on Rick Sanchez from Rick and Morty"
readme = "README.md"
authors = [
  { name = "Darren Lester (Dazbo)" },
]
requires-python = ">=3.13"
dependencies = [
    "streamlit==1.45.1",
    "google-genai==1.20.0",
    "Authlib==1.6.0",
    "google-cloud-secret-manager",
]
